requirements.txt
Flask==3.0.2
Flask-Admin==1.6.1
Flask-Login==0.6.3
Jinja2==3.1.3
numpy==1.26.4
opencv-contrib-python==4.9.0.80
pandas==2.2.0
pyparsing==3.1.1
python-dotenv==1.0.1
pyzbar==0.1.9
SQLAlchemy==2.0.25
SQLAlchemy-Utils==0.41.1
termcolor==2.4.0